Add files via upload
This project is called the capstone project. I wrote queries to analyze customer data for a subscription service with the aim of identifying segments and trends. I used basic clauses like SELECT, FROM and WHERE for subqueries. I sorted with the ORDER BY clause and also used aggregate functions like GROUP BY. I was able to extract key insights like the total number of customers in each region, the most popular subscription type and the average subscription duration for all customers to name a few. The goal is to understand customer behavior, track subscription types, and identify key trends in cancellations and renewals.